SearchModels

Provides a generic method for minimizing model errors using stochastic search, which is often used whenever the problem has no concept of derivative. This kind of problems rely on large exploration of combinatorial spaces based on error function.

SearchModels rely on basic exploration functions that can be specified for many applications.

Due to its generic specification, other kind of optimization problems can be also solved, however, it could be easier and better to use other approaches.

Using SearchModels

The code idea is to describe a configuration space (solution space) for some model, that controls how the space is explored. The model is evaluated using an error function, and the resulting errors are used to navigate the solution space.

For this purpose SearchModels expect that some methods to work. As described below.

All configuration spaces should be an specialization of the abstract type AbstractSolutionSpace.

random_configuration(space::AbstractSolutionSpace): Creates a random configuration sampling the given space

combine_configurations(a, L::AbstractVector): Combines the a solution 'a' with the list of solutions L (an array of pairs solution => error). By default it samples L for a compatible solution with a and calls combine_configurations(c1, c2). A compatible solution is defined to share the same config_type(a).

combine_configurations(c1, c2): Combines two configurations into a single one; most applications should override this method.

mutate_configuration(space::AbstractSolutionSpace, config, iter::Integer): Mutates config (a small perturbation, commonly following the space description). The iter value contains the iteration counter, it could be used to adjust the perturbation level.

mutate_configuration(space::AbstractVector, c, iter): Dispatch of the applicable mutation function whenever space defines heterogeneous types.

config_type(::T) describes the type of T, it defaults to Base.typename(T) but it could be any other value that makes sense for the application.

eltype(::AbstractSolutionSpace): Defines what kind of objects are expected to be sampled from the solution space; it requires an specialization of eltype (i.e., to add import Base: eltype).

For complex enough configurations (solutions) you may need to define specialize hash, and isequal since the search process uses it to keep a track of the evalauted configurations.

Most of the work is made in search_models function.

API

SearchModels.combine_configurations

SearchModels.combine_configurations

SearchModels.config_type

SearchModels.evaluate_queue

SearchModels.mutate_configuration

SearchModels.push_config!

SearchModels.random_configuration

SearchModels.scale

SearchModels.search_models

SearchModels.translate

SearchModels.combine_configurations — Function

combine_configurations(c1, c2)

Combines two configurations into a single one.

SearchModels.combine_configurations — Method

combine_configurations(a, L::AbstractVector)

Combines the first configuration (1st argument) with some of the given list of pairs (config => score), The a config is always at the end of L and also L is always shuffled. If you are in doubt, use the higher level interface combine_configurations(c1, c2).

SearchModels.config_type — Method

config_type(::T)

Config type identifier, it may or not be a type

SearchModels.evaluate_queue — Method

evaluate_queue(error_function, evalqueue, population, config_and_errors, parallel)

Evaluates queue of using error_function; the resulting config => errors pairs are pushed into the population.

SearchModels.mutate_configuration — Method

mutate_configuration(space::AbstractSolutionSpace, config, iter::Integer)
mutate_configuration(space::AbstractVector, c, iter)

Mutates configuration. If space is a list of spaces, then the proper space is determined.

SearchModels.push_config! — Method

push_config!(accept_config, conf, evalqueue, observed)

Pushes conf into the evaluation queue. It checks if was already observed and if it is accepted by the accept_config predicate.

SearchModels.random_configuration — Function

random_configuration(space::AbstractSolutionSpace)

Creates a random configuration sampling the given space

SearchModels.scale — Method

scale(x, s=1.1; p1=0.5, p2=0.5, lower=typemin(T), upper=typemax(T))::T

With probability p1 $x$is scaled bys; ifxis going to be scaled, then with probabilityp2x` is growth (or reduced otherwise). Minimum and maximum values can be specified.

SearchModels.search_models — Function

search_models(
    space::AbstractSolutionSpace,
    error_function::Function,  # receives only the configuration to be population
    initialpopulation=32;
    maxpopulation=initialpopulation,
    accept_config::Function=config->true,
    inspect_population::Function=(space, population) -> nothing,
    bsize=initialpopulation,
    mutbsize=4,
    crossbsize=4,
    maxiters=300,
    tol=0.001,
    verbose=true,
    parallel=:none # :none, :threads, :distributed
)

Explores the search space trying to minimize the given error function. The procedure consists on an iterative stochastic method based on an evolutionary algorithm. It is starts with initialpopulation configurations, evaluate them mutate and cross them; it selects at most maxpopulation configurations at any iteration (best ones).

space: Search space definition

error_function: the function to minimize (receives the configuration as argument)

initialpopulation: initial number of configurations to conform the population

maxpopulation: the maximum number of configurations to be kept at each iteration (based on its minimum error)

accept_config: an alternative way to deny some configuration to be evaluated (receives the configuration as argument)

inspect_population: observes the population after evaluating a beam of solutions

bsize: number of best items to be used for mutation and crossing procedures

mutbsize: number of new configurations per iteration from a mutation procedure

crossbsize: number of new configurations per iteration from a crossing procedure

maxiters: maximum iterations of the search procedure

tol: stop search tolerance to population errors on the best maxpopulation configurations; negative values will force to evaluate maxiters

verbose: controls if the verbosity of the search iterations

parallel: controls if the search is made with some kind of parallel strategy (only used for evaluation errors). Valid values are:

:none: there is no parallelization, the default value

:threads: evaluates error functions using threads

:distributed: evaluates error functions using a distributed environment (using the available workers)

SearchModels.translate — Method

translate(x::T; s=2, p1=0.5, p2=0.5, lower=typemin(T), upper=typemax(T)) where {T<:Real}

With probability p1 $x$ is modified; if $x$ is modified, then with probability p2 returns $x+s$ or $x-s$ otherwise. Minimum and maximum values can be specified.