API

SearchModels.SearchParameters

SearchModels.change

SearchModels.combine

SearchModels.combine_select

SearchModels.compatible_space

SearchModels.config_type

SearchModels.evaluate_queue

SearchModels.mutate

SearchModels.push_config!

SearchModels.scale

SearchModels.search_models

SearchModels.translate

SearchModels.SearchParameters — Type

SearchParameters(;
    maxpopulation = 32,
    bsize = maxpopulation,
    mutbsize = bsize,
    crossbsize = bsize,
    maxiters = 300,
    tol = 0.001,
    verbose = true)

Creates a mutable list of search parameters, it can be changed online via inspect_population

SearchModels.change — Method

change(x, choices; p1=0.5)

With prob p1 x is changed by some element in choices. Please note that if $x \in choices$ then the actual p1 is modified.

SearchModels.combine — Function

combine(space::AbstractSolutionSpace, c1, c2)
combine(c1, c2)

Combines two configurations into a single configuration. The three argument functions defaults to two argument function.

SearchModels.combine_select — Method

combine_select(a, L::AbstractVector)

Combines a configuration with some compatible configuation in the given list of pairs (config => score), The a config is always at the end of L and also L is always shuffled. If you are in doubt, use the higher level interface combine(c1, c2).

SearchModels.compatible_space — Method

compatible_space(space::AbstractVector, c)

Selects the compatible space for configuration c

SearchModels.config_type — Method

config_type(::T)

Config type identifier, it may or not be a type

SearchModels.evaluate_queue — Method

evaluate_queue(error_function, evalqueue, population, config_and_errors, parallel)

Evaluates queue of using error_function; the resulting config => errors pairs are pushed into the population.

SearchModels.mutate — Method

mutate(space::AbstractSolutionSpace, c, iter)
mutate(space::AbstractVector, c, iter)

Mutates configuration. If space is a list of spaces, then the proper space is determined.

SearchModels.push_config! — Method

push_config!(accept_config, conf, evalqueue, observed)

Pushes conf into the evaluation queue. It checks if was already observed and if it is accepted by the accept_config predicate.

SearchModels.scale — Method

scale(x, s=1.1; p1=0.5, p2=0.5, lower=typemin(T), upper=typemax(T))::T

With probability p1 $x$is scaled bys; ifxis going to be scaled, then with probabilityp2x` is growth (or reduced otherwise). Minimum and maximum values can be specified.

SearchModels.search_models — Function

search_models(
    space::AbstractSolutionSpace,
    error_function::Function,  # receives only the configuration to be population
    initialpopulation=32;
    maxpopulation=initialpopulation,
    accept_config::Function=config->true,
    inspect_population::Function=(space, params, population) -> nothing,
    bsize=initialpopulation,
    mutbsize=4,
    crossbsize=4,
    maxiters=300,
    tol=0.001,
    verbose=true,
    parallel=:none, # :none, :threads, :distributed,
    params = SearchParameters(;
        maxpopulation = maxpopulation,
        bsize = bsize,
        mutbsize = mutsize,
        crossbsize = crossbsize,
        maxiters = maxiters,
        tol = tol,
        verbose = verbose
    )
)

Explores the search space trying to minimize the given error function. The procedure consists on an iterative stochastic method based on an evolutionary algorithm. It is starts with initialpopulation configurations, evaluate them mutate and cross them; it selects at most maxpopulation configurations at any iteration (best ones).

space: Search space definition

error_function: the function to minimize (receives the configuration as argument)

initialpopulation: initial number of configurations to conform the population, or an initial list of configurations

maxpopulation: the maximum number of configurations to be kept at each iteration (based on its minimum error)

accept_config: an alternative way to deny some configuration to be evaluated (receives the configuration as argument)

inspect_population: observes the population after evaluating a beam of solutions

bsize: number of best items to be used for mutation and crossing procedures

mutbsize: number of new configurations per iteration from a mutation procedure

crossbsize: number of new configurations per iteration from a crossing procedure

maxiters: maximum iterations of the search procedure

tol: stop search tolerance to population errors on the best maxpopulation configurations; negative values will force to evaluate maxiters

verbose: controls if the verbosity of the search iterations

parallel: controls if the search is made with some kind of parallel strategy (only used for evaluation errors). Valid values are:

:none: there is no parallelization, the default value

:threads: evaluates error functions using threads

:distributed: evaluates error functions using a distributed environment (using the available workers)

params search parameters (created with other parameters)

SearchModels.translate — Method

translate(x::T; s=2, p1=0.5, p2=0.5, lower=typemin(T), upper=typemax(T)) where {T<:Real}

With probability p1 $x$ is modified; if $x$ is modified, then with probability p2 returns $x+s$ or $x-s$ otherwise. Minimum and maximum values can be specified.